requirements.txt
asgiref==3.8.1
attrs==23.2.0
channels==4.1.0
channels-redis==4.2.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
install==1.3.5
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
packaging==24.1
PyJWT==2.8.0
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
